% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IApre1.R
\name{IApre1}
\alias{IApre1}
\title{IApre1}
\usage{
IApre1(x)
}
\arguments{
\item{x}{a data.frame containing GeneID and gene expression of one or more patients}
}
\description{
IApre1 is used for survival prediction of stage IA lung adenocarcinoma based on a z-scale of a gene matrix
}
