% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IApre2.R
\name{IApre2}
\alias{IApre2}
\title{IApre2}
\usage{
IApre2(x)
}
\arguments{
\item{x}{a data.frame containing GeneID and gene expression of one or more patients}
}
\description{
IApre2 is used for survival prediction of stage IA lung adenocarcinoma based on a gene expression of a single patient
}
